+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Workflow modifications:
   are not correct.
 
 AQUA core complete list:
+- Refactor LRA of output and catalog entry creatro with `OutputPathBuilder` and `CatalogEntryBuilder` classes (#1932)
+- LRA cli can support realization (#1932)
 - `aqua-analysis` can accept a `--regrid` argument in order to activate the regrid on each diagnostics supporting it (#1947)
 - `--no-mount /etc/localtime` option added to the `load_aqua_container.sh` script for all HPC (#1975)
 - Upgrade to eccodes==2.41.0 (#1890)